added controllers, controller advice for handling exceptions, swagger config

+++ src/main/java/kg/parking/entity/ParkingPlace.java
@@ -24,6 +24,6 @@ public class ParkingPlace {
     private Boolean isBooked;
 
     @Column(name = "parking_fee")
-    private Double parkingFee;
+    private Long parkingFee;
 
 }
